Added compile time, _axContinue, and updated __axpus.
A compile time has been added to the JIT compiler, which prints at the end of the program's execution. Instead of exiting now, the program uses _axContinue to finish the program. This can allow for the user to have multiple programs in a single image. __axpus's red value has been changed from 255 to 200 to avoid mismatch/confusion with integers. Debug message containing the RGB values of each point has been removed.